requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.33.3
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
imageio==2.19.3
kiwisolver==1.4.3
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
sewar==0.4.5
six==1.16.0
tifffile==2022.5.4
uritemplate==4.1.1
urllib3==1.26.9
